requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
pydantic==2.9.0
httpx==0.27.0
numpy==1.26.4
python-dotenv==1.0.1
jinja2==3.1.4
